+++ code/core/src/main/java/org/betonquest/betonquest/quest/condition/redstone/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Redstone condition classes.
+ */
+package org.betonquest.betonquest.quest.condition.redstone;