+++ CHANGELOG.md
@@ -8,6 +8,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 - `drop` objective to track players dropping items
 - `statistic` placeholder to get statistic values from the player
+- `redstone` condition to require a block at a specific location to be powered to a certain level
 ### Changed
 - `lore` of `item` placeholder can now be used without number to get all lines
 - `chat` notify io now only sends messages during a conversation if `bypassInterceptor` is set to `true`